Preventing a NPE from being thrown in the cleanup code.

+++ src/test/java/org/sagebionetworks/bridge/sdk/integration/EtagsTest.java
@@ -57,7 +57,7 @@ public class EtagsTest {
         TestUser admin = TestUserHelper.getSignedInAdmin();
         SchedulesV2Api adminSchedulesApi = admin.getClient(SchedulesV2Api.class);
         
-        if (schedule != null) {
+        if (schedule != null && schedule.getGuid() != null) {
             adminSchedulesApi.deleteSchedule(schedule.getGuid()).execute();    
         }
         if (user != null) {
